change working dir for script execution, check pwd before deleting temp dir

+++ calibration/calibrate_model.sh
@@ -4,6 +4,7 @@
 ###############################################################################
 
 set -e
+cd "$(dirname "$0")"
 
 ALLOWED_DEVICES=("g2" "g3")
 
@@ -23,8 +24,13 @@ usage() {
 }
 
 cleanup_tmp() {
-    rm -rf nc_workspace 
-    rm -rf inc_tmp
+	if [[ $(pwd) == *vllm-hpu-extension/calibration ]] then
+		echo " Clearing temporary directory"
+		rm -rf nc_workspace
+		rm -rf inc_tmp
+	else
+		echo "Skipping temporary directory removal"
+	fi
 }
 create_measure_config() {
     mkdir -p $1/$2/$3
@@ -115,7 +121,7 @@ fi
 MODEL_NAME=$(extract_last_folder_name "$MODEL_PATH")
 
 echo "Step 0 - detecting used device type [g2, g3]"
-DEVICE_TYPE=$(python step-0-detect-device.py) || (echo "Detecting device process failed" && exit 1)
+DEVICE_TYPE=$(python3 step-0-detect-device.py) || (echo "Detecting device process failed" && exit 1)
 DEVICE_TYPE="g$DEVICE_TYPE"
 echo "Detected device type: $DEVICE_TYPE"
 echo "Step 0 done"
@@ -150,22 +156,22 @@ fi
 
 echo ""
 echo "1/4 Preparing calibration dataset"
-python step-1-prepare-calibration-dataset.py -m $MODEL_PATH -d $DATASET_PATH -o $MODEL_NAME $EXTRA_FLAGS || (echo "Error in step 1" && exit 1)
+python3 step-1-prepare-calibration-dataset.py -m $MODEL_PATH -d $DATASET_PATH -o $MODEL_NAME $EXTRA_FLAGS || (echo "Error in step 1" && exit 1)
 echo "Step 1/4 done"
 
 echo ""
 echo "2/4 Measuring scales"
 if $MULTI_NODE_RUN; then
-    python step-2-measure-scales.py -m $MODEL_PATH --tensor-parallel-size $TP_SIZE -d $MODEL_NAME-calibration-dataset.pkl --batch-size $BATCH_SIZE --distributed-executor-backend ray || (echo "Error in step 2" && exit 1)
+    python3 step-2-measure-scales.py -m $MODEL_PATH --tensor-parallel-size $TP_SIZE -d $MODEL_NAME-calibration-dataset.pkl --batch-size $BATCH_SIZE --distributed-executor-backend ray || (echo "Error in step 2" && exit 1)
 else
-    python step-2-measure-scales.py -m $MODEL_PATH --tensor-parallel-size $TP_SIZE -d $MODEL_NAME-calibration-dataset.pkl --batch-size $BATCH_SIZE || (echo "Error in step 2" && exit 1)
+    python3 step-2-measure-scales.py -m $MODEL_PATH --tensor-parallel-size $TP_SIZE -d $MODEL_NAME-calibration-dataset.pkl --batch-size $BATCH_SIZE || (echo "Error in step 2" && exit 1)
 fi
 
 echo "Step 2/4 done"
 
 echo ""
 echo "3/4 Postprocessing scales"
-python step-3-postprocess_measure.py -m $FP8_DIR/$MODEL_NAME/$DEVICE_TYPE/ -o inc_tmp/$MODEL_NAME/$DEVICE_TYPE/ || (echo "Error in step 3" && exit 1)
+python3 step-3-postprocess_measure.py -m $FP8_DIR/$MODEL_NAME/$DEVICE_TYPE/ -o inc_tmp/$MODEL_NAME/$DEVICE_TYPE/ || (echo "Error in step 3" && exit 1)
 cp inc_tmp/$MODEL_NAME/$DEVICE_TYPE/* $FP8_DIR/$MODEL_NAME/$DEVICE_TYPE/
 echo "Step 3/4 done"
 
@@ -179,9 +185,9 @@ fi
 echo ""
 echo "4/4 Quantize scales"
 if $MULTI_NODE_RUN; then
-    python step-4-quantize-scales.py --model $MODEL_PATH --tensor-parallel-size $TP_SIZE --distributed-executor-backend ray || (echo "Error in step 4" && exit 1)
+    python3 step-4-quantize-scales.py --model $MODEL_PATH --tensor-parallel-size $TP_SIZE --distributed-executor-backend ray || (echo "Error in step 4" && exit 1)
 else
-    python step-4-quantize-scales.py --model $MODEL_PATH --tensor-parallel-size $TP_SIZE || (echo "Error in step 4" && exit 1)
+    python3 step-4-quantize-scales.py --model $MODEL_PATH --tensor-parallel-size $TP_SIZE || (echo "Error in step 4" && exit 1)
 fi
 
 cleanup_tmp
